etape_13_deployable/requirements.txt
# ── API Framework ─────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# ── LLM & RAG ─────────────────────────────────────────────────────────────
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
openai>=1.0.0
chromadb>=0.4.0          # Base vectorielle (optionnel — app fonctionne sans)

# ── Sécurité ──────────────────────────────────────────────────────────────
slowapi>=0.1.9
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1

# ── Monitoring ────────────────────────────────────────────────────────────
prometheus-client>=0.19.0
psutil>=5.9.0

# ── Persistance ───────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.0.0

# ── HTTP ──────────────────────────────────────────────────────────────────
httpx>=0.27.0
